+++ packages/xen-api/src/index.js
@@ -7,8 +7,8 @@ import { BaseError } from 'make-error'
 import { EventEmitter } from 'events'
 import { fibonacci } from 'iterable-backoff'
 import {
-  filter,
   forEach,
+  forOwn,
   isArray,
   isInteger,
   map,
@@ -37,7 +37,7 @@ const debug = createDebug('xen-api')
 
 // ===================================================================
 
-// in seconds
+// in seconds!
 const EVENT_TIMEOUT = 60
 
 // http://www.gnu.org/software/libc/manual/html_node/Error-Codes.html
@@ -271,17 +271,14 @@ export class Xapi extends EventEmitter {
     super()
 
     this._allowUnauthorized = opts.allowUnauthorized
-    this._auth = opts.auth
     this._callTimeout = makeCallSetting(opts.callTimeout, 0)
-    this._debounce = opts.debounce == null ? 200 : opts.debounce
     this._pool = null
     this._readOnly = Boolean(opts.readOnly)
     this._RecordsByType = createObject(null)
     this._sessionId = null
-    ;(this._objects = new Collection()).getKey = getKey
-    ;(this._objectsByRef = createObject(null))[NULL_REF] = undefined
-    const url = (this._url = parseUrl(opts.url))
 
+    this._auth = opts.auth
+    const url = (this._url = parseUrl(opts.url))
     if (this._auth === undefined) {
       const user = url.username
       if (user !== undefined) {
@@ -294,7 +291,19 @@ export class Xapi extends EventEmitter {
       }
     }
 
-    if (opts.watchEvents !== false) {
+    ;(this._objects = new Collection()).getKey = getKey
+    this._debounce = opts.debounce == null ? 200 : opts.debounce
+    this._watchedTypes = undefined
+    this._watching = false
+
+    this.on(DISCONNECTED, this._clearObjects)
+    this._clearObjects()
+
+    const { watchEvents } = opts
+    if (watchEvents !== false) {
+      if (Array.isArray(watchEvents)) {
+        this._watchedTypes = watchEvents
+      }
       this.watchEvents()
     }
   }
@@ -302,19 +311,14 @@ export class Xapi extends EventEmitter {
   watchEvents() {
     this._eventWatchers = createObject(null)
 
-    this._fromToken = ''
-
-    this._nTasks = 0
-
     this._taskWatchers = Object.create(null)
 
     if (this.status === CONNECTED) {
-      this._watchEvents()
+      this._watchEventsWrapper()
     }
 
-    this.on('connected', this._watchEvents)
+    this.on('connected', this._watchEventsWrapper)
     this.on('disconnected', () => {
-      this._fromToken = ''
       this._objects.clear()
     })
   }
@@ -403,42 +407,55 @@ export class Xapi extends EventEmitter {
     )
   }
 
-  connect() {
+  async connect() {
     const { status } = this
 
     if (status === CONNECTED) {
-      return Promise.reject(new Error('already connected'))
+      throw new Error('already connected')
     }
 
     if (status === CONNECTING) {
-      return Promise.reject(new Error('already connecting'))
+      throw new Error('already connecting')
     }
 
     const auth = this._auth
     if (auth === undefined) {
-      return Promise.reject(new Error('missing credentials'))
+      throw new Error('missing credentials')
     }
 
     this._sessionId = CONNECTING
 
-    return this._transportCall('session.login_with_password', [
-      auth.user,
-      auth.password,
-    ]).then(
-      async sessionId => {
-        this._sessionId = sessionId
-        this._pool = (await this.getAllRecords('pool'))[0]
+    try {
+      const [methods, sessionId] = await Promise.all([
+        this._transportCall('system.listMethods', []),
+        this._transportCall('session.login_with_password', [
+          auth.user,
+          auth.password,
+        ]),
+      ])
 
-        debug('%s: connected', this._humanId)
+      // Uses introspection to list available types.
+      const types = (this._types = methods
+        .filter(isGetAllRecordsMethod)
+        .map(method => method.slice(0, method.indexOf('.'))))
+      this._lcToTypes = { __proto__: null }
+      types.forEach(type => {
+        const lcType = type.toLowerCase()
+        if (lcType !== type) {
+          this._lcToTypes[lcType] = type
+        }
+      })
 
-        this.emit(CONNECTED)
-      },
-      error => {
-        this._sessionId = null
+      this._sessionId = sessionId
+      this._pool = (await this.getAllRecords('pool'))[0]
 
-        throw error
-      }
-    )
+      debug('%s: connected', this._humanId)
+      this.emit(CONNECTED)
+    } catch (error) {
+      this._sessionId = null
+
+      throw error
+    }
   }
 
   disconnect() {
@@ -501,6 +518,10 @@ export class Xapi extends EventEmitter {
     return promise
   }
 
+  getField(type, ref, field) {
+    return this._sessionCall(`${type}.get_${field}`, [ref])
+  }
+
   // Nice getter which returns the object for a given $id (internal to
   // this lib), UUID (unique identifier that some objects have) or
   // opaque reference (internal to XAPI).
@@ -552,6 +573,10 @@ export class Xapi extends EventEmitter {
     )
   }
 
+  getRecords(type, refs) {
+    return Promise.all(refs.map(ref => this.getRecord(type, ref)))
+  }
+
   async getAllRecords(type) {
     return map(
       await this._sessionCall(`${type}.get_all_records`),
@@ -567,7 +592,7 @@ export class Xapi extends EventEmitter {
   }
 
   @cancelable
-  getResource($cancelToken, pathname, { host, query, task }) {
+  getResource($cancelToken, pathname, { host, query, task } = {}) {
     return this._autoTask(task, `Xapi#getResource ${pathname}`).then(
       taskRef => {
         query = { ...query, session_id: this.sessionId }
@@ -720,39 +745,36 @@ export class Xapi extends EventEmitter {
     )
   }
 
-  setField({ $type, $ref }, field, value) {
-    return this.call(`${$type}.set_${field}`, $ref, value).then(noop)
+  setField(type, ref, field, value) {
+    return this.call(`${type}.set_${field}`, ref, value).then(noop)
   }
 
-  setFieldEntries(record, field, entries) {
+  setFieldEntries(type, ref, field, entries) {
     return Promise.all(
       getKeys(entries).map(entry => {
         const value = entries[entry]
         if (value !== undefined) {
-          return value === null
-            ? this.unsetFieldEntry(record, field, entry)
-            : this.setFieldEntry(record, field, entry, value)
+          return this.setFieldEntry(type, ref, field, entry, value)
         }
       })
     ).then(noop)
   }
 
-  async setFieldEntry({ $type, $ref }, field, entry, value) {
+  async setFieldEntry(type, ref, field, entry, value) {
+    if (value === null) {
+      return this.call(`${type}.remove_from_${field}`, ref, entry).then(noop)
+    }
     while (true) {
       try {
-        await this.call(`${$type}.add_to_${field}`, $ref, entry, value)
+        await this.call(`${type}.add_to_${field}`, ref, entry, value)
         return
       } catch (error) {
         if (error == null || error.code !== 'MAP_DUPLICATE_KEY') {
           throw error
         }
       }
-      await this.unsetFieldEntry({ $type, $ref }, field, entry)
+      await this.call(`${type}.remove_from_${field}`, ref, entry)
     }
-  }
-
-  unsetFieldEntry({ $type, $ref }, field, entry) {
-    return this.call(`${$type}.remove_from_${field}`, $ref, entry)
   }
 
   watchTask(ref) {
@@ -788,6 +810,15 @@ export class Xapi extends EventEmitter {
     return this._objects
   }
 
+  _clearObjects() {
+    ;(this._objectsByRef = createObject(null))[NULL_REF] = undefined
+    this._nTasks = 0
+    this._objects.clear()
+    this.objectsFetched = new Promise(resolve => {
+      this._resolveObjectsFetched = resolve
+    })
+  }
+
   // return a promise which resolves to a task ref or undefined
   _autoTask(task = this._taskWatchers !== undefined, name) {
     if (task === false) {
@@ -803,7 +834,7 @@ export class Xapi extends EventEmitter {
   }
 
   // Medium level call: handle session errors.
-  _sessionCall(method, args) {
+  _sessionCall(method, args, timeout = this._callTimeout(method, args)) {
     try {
       if (startsWith(method, 'session.')) {
         throw new Error('session.*() methods are disabled from this interface')
@@ -827,7 +858,7 @@ export class Xapi extends EventEmitter {
             return this.connect().then(() => this._sessionCall(method, args))
           }
         ),
-        this._callTimeout(method, args)
+        timeout
       )
     } catch (error) {
       return Promise.reject(error)
@@ -906,7 +937,12 @@ export class Xapi extends EventEmitter {
 
   _processEvents(events) {
     forEach(events, event => {
-      const { class: type, ref } = event
+      let type = event.class
+      const lcToTypes = this._lcToTypes
+      if (type in lcToTypes) {
+        type = lcToTypes[type]
+      }
+      const { ref } = event
       if (event.operation === 'del') {
         this._removeObject(type, ref)
       } else {
@@ -915,34 +951,112 @@ export class Xapi extends EventEmitter {
     })
   }
 
-  _watchEvents() {
-    const loop = () =>
-      this.status === CONNECTED &&
-      pTimeout
-        .call(
-          this._sessionCall('event.from', [
-            ['*'],
-            this._fromToken,
-            EVENT_TIMEOUT + 0.1, // Force float.
-          ]),
-          EVENT_TIMEOUT * 1.1e3 // 10% longer than the XenAPI timeout
+  // - prevent multiple watches
+  // - swallow errors
+  async _watchEventsWrapper() {
+    if (!this._watching) {
+      this._watching = true
+      try {
+        await this._watchEvents()
+      } catch (error) {
+        console.error('_watchEventsWrapper', error)
+      }
+      this._watching = false
+    }
+  }
+
+  // TODO: cancelation
+  async _watchEvents() {
+    this._clearObjects()
+
+    // compute the initial token for the event loop
+    //
+    // we need to do this before the initial fetch to avoid losing events
+    let fromToken
+    try {
+      fromToken = await this._sessionCall('event.inject', [
+        'pool',
+        this._pool.$ref,
+      ])
+    } catch (error) {
+      if (isMethodUnknown(error)) {
+        return this._watchEventsLegacy()
+      }
+    }
+
+    const types = this._watchedTypes || this._types
+
+    // initial fetch
+    const flush = this.objects.bufferEvents()
+    try {
+      await Promise.all(
+        types.map(async type => {
+          try {
+            // FIXME: use _transportCall to avoid auto-reconnection
+            forOwn(
+              await this._sessionCall(`${type}.get_all_records`),
+              (record, ref) => {
+                // we can bypass _processEvents here because they are all *add*
+                // event and all objects are of the same type
+                this._addObject(type, ref, record)
+              }
+            )
+          } catch (error) {
+            // there is nothing ideal to do here, do not interrupt event
+            // handling
+            if (error != null && error.code !== 'MESSAGE_REMOVED') {
+              console.warn('_watchEvents', 'initial fetch', type, error)
+            }
+          }
+        })
+      )
+    } finally {
+      flush()
+    }
+    this._resolveObjectsFetched()
+
+    // event loop
+    const debounce = this._debounce
+    while (true) {
+      if (debounce != null) {
+        await pDelay(debounce)
+      }
+
+      let result
+      try {
+        result = await this._sessionCall(
+          'event.from',
+          [
+            types,
+            fromToken,
+            EVENT_TIMEOUT + 0.1, // must be float for XML-RPC transport
+          ],
+          EVENT_TIMEOUT * 1e3 * 1.1
         )
-        .then(onSuccess, onFailure)
+      } catch (error) {
+        if (error instanceof TimeoutError) {
+          continue
+        }
+        if (areEventsLost(error)) {
+          return this._watchEvents()
+        }
+        throw error
+      }
 
-    const onSuccess = ({ events, token, valid_ref_counts: { task } }) => {
-      this._fromToken = token
-      this._processEvents(events)
+      fromToken = result.token
+      this._processEvents(result.events)
 
-      if (task !== this._nTasks) {
-        this._sessionCall('task.get_all_records')
-          .then(tasks => {
+      // detect and fix disappearing tasks (e.g. when toolstack restarts)
+      if (result.valid_ref_counts.task !== this._nTasks) {
+        await ignoreErrors.call(
+          this._sessionCall('task.get_all_records').then(tasks => {
             const toRemove = new Set()
-            forEach(this.objects.all, object => {
+            forOwn(this.objects.all, object => {
               if (object.$type === 'task') {
                 toRemove.add(object.$ref)
               }
             })
-            forEach(tasks, (task, ref) => {
+            forOwn(tasks, (task, ref) => {
               toRemove.delete(ref)
               this._addObject('task', ref, task)
             })
@@ -950,40 +1064,9 @@ export class Xapi extends EventEmitter {
               this._removeObject('task', ref)
             })
           })
-          .catch(noop)
+        )
       }
-
-      const debounce = this._debounce
-      return debounce != null ? pDelay(debounce).then(loop) : loop()
     }
-    const onFailure = error => {
-      if (error instanceof TimeoutError) {
-        return loop()
-      }
-
-      if (areEventsLost(error)) {
-        this._fromToken = ''
-        this._objects.clear()
-
-        return loop()
-      }
-
-      throw error
-    }
-
-    ignoreErrors.call(
-      pCatch.call(
-        loop(),
-        isMethodUnknown,
-
-        // If the server failed, it is probably due to an excessively
-        // large response.
-        // Falling back to legacy events watch should be enough.
-        error => error && error.res && error.res.statusCode === 500,
-
-        () => this._watchEventsLegacy()
-      )
-    )
   }
 
   // This method watches events using the legacy `event.next` XAPI
@@ -991,17 +1074,13 @@ export class Xapi extends EventEmitter {
   //
   // It also has to manually get all objects first.
   _watchEventsLegacy() {
-    const getAllObjects = () => {
-      return this._sessionCall('system.listMethods').then(methods => {
-        // Uses introspection to determine the methods to use to get
-        // all objects.
-        const getAllRecordsMethods = filter(methods, isGetAllRecordsMethod)
-
-        return Promise.all(
-          map(getAllRecordsMethods, method =>
-            this._sessionCall(method).then(
+    const getAllObjects = async () => {
+      const flush = this.objects.bufferEvents()
+      try {
+        await Promise.all(
+          this._types.map(type =>
+            this._sessionCall(`${type}.get_all_records`).then(
               objects => {
-                const type = method.slice(0, method.indexOf('.')).toLowerCase()
                 forEach(objects, (object, ref) => {
                   this._addObject(type, ref, object)
                 })
@@ -1014,7 +1093,10 @@ export class Xapi extends EventEmitter {
             )
           )
         )
-      })
+      } finally {
+        flush()
+      }
+      this._resolveObjectsFetched()
     }
 
     const watchEvents = () =>
@@ -1050,8 +1132,7 @@ export class Xapi extends EventEmitter {
       const nFields = fields.length
       const xapi = this
 
-      const objectsByRef = this._objectsByRef
-      const getObjectByRef = ref => objectsByRef[ref]
+      const getObjectByRef = ref => this._objectsByRef[ref]
 
       Record = function(ref, data) {
         defineProperties(this, {
@@ -1069,7 +1150,7 @@ export class Xapi extends EventEmitter {
       const props = { $type: type }
       fields.forEach(field => {
         props[`set_${field}`] = function(value) {
-          return xapi.setField(this, field, value)
+          return xapi.setField(this.$type, this.$ref, field, value)
         }
 
         const $field = (field in RESERVED_FIELDS ? '$$' : '$') + field
@@ -1093,19 +1174,21 @@ export class Xapi extends EventEmitter {
             const value = this[field]
             const result = {}
             getKeys(value).forEach(key => {
-              result[key] = objectsByRef[value[key]]
+              result[key] = xapi._objectsByRef[value[key]]
             })
             return result
           }
-          props[`update_${field}`] = function(entries) {
-            return xapi.setFieldEntries(this, field, entries)
+          props[`update_${field}`] = function(entries, value) {
+            return typeof entries === 'string'
+              ? xapi.setFieldEntry(this.$type, this.$ref, field, entries, value)
+              : xapi.setFieldEntries(this.$type, this.$ref, field, entries)
           }
         } else if (value === '' || isOpaqueRef(value)) {
           // 2019-02-07 - JFT: even if `value` should not be an empty string for
           // a ref property, an user had the case on XenServer 7.0 on the CD VBD
           // of a VM created by XenCenter
           getters[$field] = function() {
-            return objectsByRef[this[field]]
+            return xapi._objectsByRef[this[field]]
           }
         }
       })
@@ -1141,9 +1224,15 @@ Xapi.prototype._transportCall = reduce(
             error = wrapError(error)
           }
 
+          // do not log the session ID
+          //
+          // TODO: should log at the session level to avoid logging sensitive
+          // values?
+          const params = args[0] === this._sessionId ? args.slice(1) : args
+
           error.call = {
             method,
-            params: replaceSensitiveValues(args, '* obfuscated *'),
+            params: replaceSensitiveValues(params, '* obfuscated *'),
           }
           throw error
         })