+++ packages/xo-web/src/xo-app/logs/backup-ng/index.js
@@ -13,7 +13,11 @@ import { createGetObjectsOfType } from 'selectors'
 import { get } from '@xen-orchestra/defined'
 import { injectState, provideState } from 'reaclette'
 import { isEmpty, filter, map, keyBy } from 'lodash'
-import { subscribeBackupNgJobs, subscribeBackupNgLogs } from 'xo'
+import {
+  subscribeBackupNgJobs,
+  subscribeBackupNgLogs,
+  subscribeMetadataBackupJobs,
+} from 'xo'
 
 import LogAlertBody from './log-alert-body'
 import LogAlertHeader from './log-alert-header'
@@ -34,6 +38,7 @@ export const LogStatus = ({ log, tooltip = _('logDisplayDetails') }) => {
   return (
     <ActionButton
       btnStyle={className}
+      disabled={log.status !== 'failure' && isEmpty(log.tasks)}
       handler={showTasks}
       handlerParam={log.id}
       icon='preview'
@@ -84,8 +89,8 @@ const COLUMNS = [
   },
   {
     name: _('labelSize'),
-    itemRenderer: ({ tasks: vmTasks }) => {
-      if (isEmpty(vmTasks)) {
+    itemRenderer: ({ tasks: vmTasks, jobId }, { jobs }) => {
+      if (get(() => jobs[jobId].type) !== 'backup' || isEmpty(vmTasks)) {
         return null
       }
 
@@ -151,6 +156,8 @@ export default decorate([
         cb(logs && filter(logs, log => log.message !== 'restore'))
       ),
     jobs: cb => subscribeBackupNgJobs(jobs => cb(keyBy(jobs, 'id'))),
+    metadataJobs: cb =>
+      subscribeMetadataBackupJobs(jobs => cb(keyBy(jobs, 'id'))),
   }),
   provideState({
     computed: {
@@ -167,6 +174,7 @@ export default decorate([
               }
             : log
         ),
+      jobs: (_, { jobs, metadataJobs }) => ({ ...jobs, ...metadataJobs }),
     },
   }),
   injectState,
@@ -180,7 +188,7 @@ export default decorate([
           collection={state.logs}
           columns={COLUMNS}
           component={SortedTable}
-          data-jobs={jobs}
+          data-jobs={state.jobs}
           emptyMessage={_('noLogs')}
           filters={LOG_FILTERS}
         />